fixed MSVC warning
Former-commit-id: 890d7d39ceb814838c6a85cd32712e5f93ff0c73

+++ src/Codecs/QualCodec.cc
@@ -261,7 +261,7 @@ void QualEncoder::loadFastaReference(const std::string &rname)
             foundFastaReference = true;
             reference = fastaReference.sequence;
             referencePosMin = 0;
-            referencePosMax = reference.size() - 1;
+            referencePosMax = (uint32_t)reference.size() - 1;
         }
     }
